Move unused constabulary views to separate module.
Original feature commit:
  - ICMSLST-2393 - View for constabularies to see issued documents.
